requirements.txt
torch
torchvision
pyyaml
tqdm
pandas
matplotlib